refactor: centralize image decode result flow
